Renaming api to client
