+++ lib/routes/index.js
@@ -8,4 +8,4 @@ r.use(require('./invalidate'))
 r.use(require('./tarballs'))
 r.use(require('./packages'))
 r.use(require('./publish'))
-r.use(require('./api'))
+r.use(require('./client'))
